orgsetting refactor
