+++ docs/docs/auto-docs/components/OrgSettings/ActionItemCategories/Modal/CategoryModal/interfaces/InterfaceActionItemCategoryModal.md
@@ -1,0 +1,72 @@
+[Admin Docs](/)
+
+***
+
+# Interface: InterfaceActionItemCategoryModal
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:26](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L26)
+
+Props for the `CategoryModal` component.
+
+isOpen - The state of the modal.
+hide - The function to hide the modal.
+refetchCategories - The function to refetch the categories.
+orgId - The organization ID.
+category - The category to be edited.
+mode - The mode of the modal.
+
+## Properties
+
+### category
+
+> **category**: [`InterfaceActionItemCategoryInfo`](../../../../../../utils/interfaces/interfaces/InterfaceActionItemCategoryInfo.md)
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:31](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L31)
+
+***
+
+### hide()
+
+> **hide**: () => `void`
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:28](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L28)
+
+#### Returns
+
+`void`
+
+***
+
+### isOpen
+
+> **isOpen**: `boolean`
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:27](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L27)
+
+***
+
+### mode
+
+> **mode**: `"create"` \| `"edit"`
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:32](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L32)
+
+***
+
+### orgId
+
+> **orgId**: `string`
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:30](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L30)
+
+***
+
+### refetchCategories()
+
+> **refetchCategories**: () => `void`
+
+Defined in: [src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx:29](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/ActionItemCategories/Modal/CategoryModal.tsx#L29)
+
+#### Returns
+
+`void`
