+++ docs/docs/auto-docs/components/OrgSettings/AgendaItemCategories/Preview/AgendaCategoryPreviewModal/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`props`, `deprecatedLegacyContext`?): `ReactNode`
 
-Defined in: [src/components/OrgSettings/AgendaItemCategories/AgendaCategoryPreviewModal.tsx:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/AgendaItemCategories/AgendaCategoryPreviewModal.tsx#L38)
+Defined in: [src/components/OrgSettings/AgendaItemCategories/Preview/AgendaCategoryPreviewModal.tsx:38](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/AgendaItemCategories/Preview/AgendaCategoryPreviewModal.tsx#L38)
 
 AgendaCategoryPreviewModal component is used to preview the agenda category details like name, description, createdBy
