+++ docs/docs/auto-docs/components/OrgSettings/AgendaItemCategories/Update/AgendaCategoryUpdateModal/functions/default.md
@@ -6,7 +6,7 @@
 
 > **default**(`props`, `deprecatedLegacyContext`?): `ReactNode`
 
-Defined in: [src/components/OrgSettings/AgendaItemCategories/AgendaCategoryUpdateModal.tsx:40](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/AgendaItemCategories/AgendaCategoryUpdateModal.tsx#L40)
+Defined in: [src/components/OrgSettings/AgendaItemCategories/Update/AgendaCategoryUpdateModal.tsx:40](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/components/OrgSettings/AgendaItemCategories/Update/AgendaCategoryUpdateModal.tsx#L40)
 
 AgendaCategoryUpdateModal component is used to update the agenda category details like name, description
