+++ src/components/AgendaCategory/AgendaCategoryContainer.tsx
@@ -12,9 +12,9 @@ import {
 import type { InterfaceAgendaItemCategoryInfo } from 'utils/interfaces';
 import styles from '../../style/app.module.css';
 
-import AgendaCategoryDeleteModal from 'components/OrgSettings/AgendaItemCategories/AgendaCategoryDeleteModal';
-import AgendaCategoryPreviewModal from 'components/OrgSettings/AgendaItemCategories/AgendaCategoryPreviewModal';
-import AgendaCategoryUpdateModal from 'components/OrgSettings/AgendaItemCategories/AgendaCategoryUpdateModal';
+import AgendaCategoryDeleteModal from 'components/OrgSettings/AgendaItemCategories/Delete/AgendaCategoryDeleteModal';
+import AgendaCategoryPreviewModal from 'components/OrgSettings/AgendaItemCategories/Preview/AgendaCategoryPreviewModal';
+import AgendaCategoryUpdateModal from 'components/OrgSettings/AgendaItemCategories/Update/AgendaCategoryUpdateModal';
 
 /**
  * Component for displaying and managing agenda item categories.
