+++ src/components/OrgSettings/AgendaItemCategories/OrganizationAgendaCategory.tsx
@@ -12,7 +12,7 @@ import { CREATE_AGENDA_ITEM_CATEGORY_MUTATION } from 'GraphQl/Mutations/mutation
 
 import type { InterfaceAgendaItemCategoryList } from 'utils/interfaces';
 import AgendaCategoryContainer from 'components/AgendaCategory/AgendaCategoryContainer';
-import AgendaCategoryCreateModal from './AgendaCategoryCreateModal';
+import AgendaCategoryCreateModal from './Create/AgendaCategoryCreateModal';
 import styles from 'style/app.module.css';
 import Loader from 'components/Loader/Loader';
 import SearchBar from 'subComponents/SearchBar';
